+++ packages/rum/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./config";
 export * from "./utils";
+export * from "./types";
